requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
jinja2==3.1.6
pymupdf==1.25.5
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.2.1
numpy==2.2.5
faiss-cpu==1.10.0
sentence-transformers==4.1.0
transformers==4.51.3
torch==2.6.0
rapidfuzz==3.13.0
python-magic==0.4.27
watchdog==6.0.0
aiofiles==24.1.0
